moving scene files to gameLogic
