+++ src/gameLogic/levels/Two/index.ts
@@ -24,7 +24,7 @@ import {
   V_DESK_SIZE,
   DX,
   DY,
-} from '../../index';
+} from '../../../constants';
 
 //ORIGEM DO JOGADOR E DO CHÃO//////////////////////////////////////////////////////////
 const mapOrigin = {
